Moved the layout reducer ready to add an auto layout handler
