+++ src/malcolm/reducer/layout/layout.reducer.test.js
