+++ src/malcolm/reducer/malcolmReducer.test.js
@@ -15,7 +15,7 @@ import {
 import { AlarmStates } from '../../malcolmWidgets/attributeDetails/attributeAlarm/attributeAlarm.component';
 import NavigationReducer from './navigation.reducer';
 import MethodReducer from './method.reducer';
-import LayoutReducer, { LayoutReduxReducer } from './layout.reducer';
+import LayoutReducer, { LayoutReduxReducer } from './layout/layout.reducer';
 
 jest.mock('./navigation.reducer');
 jest.mock('./method.reducer');
